Move all files in models/helpers to expungeservice root
Originally the thought was these files do actions on the models but aren't models themselves so they should be in a sub-folder in models. But now I am thinking that most of these files drive some core logic of the backend, and so these files should just all be at the expungeservice root.
