+++ src/backend/tests/factories/charge_factory.py
@@ -2,7 +2,7 @@ from datetime import date as date_class
 
 from expungeservice.models.ambiguous import AmbiguousCharge
 from expungeservice.models.charge import Charge
-from expungeservice.models.helpers.charge_creator import ChargeCreator
+from expungeservice.charge_creator import ChargeCreator
 from expungeservice.models.disposition import Disposition
