+++ src/backend/tests/models/charge_types/test_manufacture_delivery.py
@@ -1,5 +1,5 @@
 from expungeservice.models.expungement_result import EligibilityStatus
-from expungeservice.models.helpers.record_merger import RecordMerger
+from expungeservice.record_merger import RecordMerger
 from tests.factories.charge_factory import ChargeFactory
 from tests.models.test_charge import Dispositions
