+++ src/backend/tests/models/charge_types/test_person_felony.py
@@ -1,7 +1,6 @@
 from expungeservice.models.expungement_result import EligibilityStatus
 from expungeservice.models.charge_types.person_felony import PersonFelonyClassB
-from expungeservice.models.charge_types.felony_class_c import FelonyClassC
-from expungeservice.models.helpers.record_merger import RecordMerger
+from expungeservice.record_merger import RecordMerger
 from tests.factories.charge_factory import ChargeFactory
 from tests.models.test_charge import Dispositions
 import pytest
