+++ src/backend/tests/models/test_record_summarizer.py
@@ -1,6 +1,6 @@
 from expungeservice.models.disposition import Disposition
-from expungeservice.models.helpers.record_merger import RecordMerger
-from expungeservice.models.helpers.record_summarizer import RecordSummarizer
+from expungeservice.record_merger import RecordMerger
+from expungeservice.record_summarizer import RecordSummarizer
 from expungeservice.expunger import Expunger
 from expungeservice.models.record import Record
 from tests.factories.case_factory import CaseFactory
